deploy/requirements.txt
Flask-Assets
Flask
pyScss
cssmin
Flask-HTTPAuth
gunicorn
py-bcrypt
psutil